+++ docs/examples/desktop/desktop.js
@@ -83,6 +83,8 @@ var w_titles = [
 function createWindow(){
     var index = Metro.utils.random(0, 3);
     Desktop.createWindow({
+        resizeable: true,
+        draggable: true,
         width: 300,
         icon: "<span class='mif-"+w_icons[index]+"'></span>",
         title: w_titles[index],
@@ -90,8 +92,40 @@ function createWindow(){
     });
 }
 
+function createWindowWithCustomButtons(){
+    var index = Metro.utils.random(0, 3);
+    var customButtons = [
+        {
+            html: "<span class='mif-rocket'></span>",
+            cls: "sys-button",
+            onclick: "alert('You press user button')"
+        },
+        {
+            html: "<span class='mif-user'></span>",
+            cls: "alert",
+            onclick: "alert('You press user button')"
+        },
+        {
+            html: "<span class='mif-cog'></span>",
+            cls: "warning",
+            onclick: "alert('You press cog button')"
+        }
+    ];
+    Desktop.createWindow({
+        resizeable: true,
+        draggable: true,
+        customButtons: customButtons,
+        width: 360,
+        icon: "<span class='mif-"+w_icons[index]+"'></span>",
+        title: w_titles[index],
+        content: "<div class='p-2'>This is desktop demo created with Metro 4 Components Library.<br><br>This window has a custom buttons in caption.</div>"
+    });
+}
+
 function createWindowModal(){
     Desktop.createWindow({
+        resizeable: false,
+        draggable: true,
         width: 300,
         icon: "<span class='mif-cogs'></span>",
         title: "Modal window",
@@ -114,6 +148,8 @@ function createWindowModal(){
 
 function createWindowYoutube(){
     Desktop.createWindow({
+        resizeable: true,
+        draggable: true,
         width: 500,
         icon: "<span class='mif-youtube'></span>",
         title: "Youtube video",